refacoring of window.js etc
