+++ gear-sim-window/actions/invalid-action.js
@@ -1,0 +1,9 @@
+module.exports = class SelectFolderAction {
+    constructor(url) {
+        this.url = url;
+    }
+
+    process() {
+        return { error: 'Invalid url ' + this.url };
+    }
+}
