+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "dngearsim-desktop",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "description": "Dragon Nest Gear Simulator - Desktop Edition",
   "productName": "dngearsim-desktop",
   "main": "main.js",
